requirements.txt
mysqlclient==1.3.9
Django==1.8.4
django-extensions==1.7.7
djangorestframework==3.3.2
uWSGI==2.0.12
serpy==0.1.0
six==1.10.0
appdirs==1.4.3
configparser==3.5.0
packaging==16.8
pyparsing==2.2.0

